# docs/openapi.yaml
openapi: 3.0.4
info:
  title: Battleship Game API
  version: '1.0'
paths:
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Registers a new user identity and creates their game profile in one step.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/auth/signin:
    post:
      tags:
        - Auth
      summary: Signs in an existing user and returns access and refresh tokens.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/SignInRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/auth/refresh:
    post:
      tags:
        - Auth
      summary: Exchanges a refresh token for a new token pair.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokenResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/auth/logout:
    post:
      tags:
        - Auth
      summary: Invalidates the session and signs the user out.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '204':
          description: No Content
  /api/games:
    post:
      tags:
        - Games
      summary: Creates a new game.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '201':
          description: Game successfully created.
          content:
            text/plain:
              schema:
                type: string
                format: uuid
            application/json:
              schema:
                type: string
                format: uuid
            text/json:
              schema:
                type: string
                format: uuid
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/games/{id}':
    get:
      tags:
        - Games
      summary: Retrieves a game by ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Returns a game.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GetGameQueryResult'
            application/json:
              schema:
                $ref: '#/components/schemas/GetGameQueryResult'
            text/json:
              schema:
                $ref: '#/components/schemas/GetGameQueryResult'
        '400':
          description: Invalid input data.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Game not found.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '500':
          description: Internal server error.
  /api/games/active:
    get:
      tags:
        - Games
      summary: 'Returns the caller''s active game, or 204 No Content when none is in progress.'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GetGameQueryResult'
            application/json:
              schema:
                $ref: '#/components/schemas/GetGameQueryResult'
            text/json:
              schema:
                $ref: '#/components/schemas/GetGameQueryResult'
        '204':
          description: No Content
        '403':
          description: Forbidden
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/games/{id}/ships':
    post:
      tags:
        - Games
      summary: Adds a ship to a certain board side.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceShipRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/PlaceShipRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/PlaceShipRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                type: string
                format: uuid
            application/json:
              schema:
                type: string
                format: uuid
            text/json:
              schema:
                type: string
                format: uuid
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/games/{id}/attacks':
    post:
      tags:
        - Games
      summary: Attacks a cell on a certain board side.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttackRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/AttackRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/AttackRequest'
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/LastRoundResult'
            application/json:
              schema:
                $ref: '#/components/schemas/LastRoundResult'
            text/json:
              schema:
                $ref: '#/components/schemas/LastRoundResult'
        '400':
          description: Bad Request
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/games/{id}/state':
    put:
      tags:
        - Games
      summary: 'Updates game state (e.g., transitions from Ready to Started).'
      description: "This endpoint manages game state transitions. Currently supports transitioning from Ready to Started state.\r\n            \r\nUsage:\r\nPUT /api/games/{id}/state\r\n{ \"state\": \"started\" }\r\n            \r\nValid transitions:\r\n- \"started\": Transitions game from Ready state to Started state (allows attacks to begin)"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateGameStateRequest'
          text/json:
            schema:
              $ref: '#/components/schemas/UpdateGameStateRequest'
          application/*+json:
            schema:
              $ref: '#/components/schemas/UpdateGameStateRequest'
      responses:
        '200':
          description: Game state successfully updated.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
        '400':
          description: Invalid state transition or invalid state value.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '404':
          description: Game not found.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '409':
          description: Invalid current state for requested transition.
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    get:
      tags:
        - Games
      summary: Retrieves a game state.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
            application/json:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
            text/json:
              schema:
                $ref: '#/components/schemas/GameStateResponse'
        '404':
          description: Not Found
          content:
            text/plain:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
            text/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /api/players/me:
    get:
      tags:
        - Players
      summary: Returns the authenticated caller's own player profile.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/players/{id}':
    get:
      tags:
        - Players
      summary: Gets a player by ID.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  '/api/players/{username}':
    get:
      tags:
        - Players
      summary: Gets a player by username.
      parameters:
        - name: username
          in: path
          required: true
          schema:
            pattern: '^[A-Za-z]*$'
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerResponse'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
components:
  schemas:
    AttackRequest:
      type: object
      properties:
        cell:
          type: string
          nullable: true
      additionalProperties: false
    AuthTokenResponse:
      type: object
      properties:
        accessToken:
          type: string
          nullable: true
        refreshToken:
          type: string
          nullable: true
        expiresInSeconds:
          type: integer
          format: int32
      additionalProperties: false
    BoardSide:
      enum:
        - None
        - Player
        - Opponent
      type: string
    CellState:
      enum:
        - None
        - Clear
        - Occupied
        - Hit
        - Missed
      type: string
    CreateGameRequest:
      type: object
      properties:
        boardSize:
          type: integer
          format: int32
          nullable: true
        opponentStrategy:
          $ref: '#/components/schemas/OpponentStrategy'
      additionalProperties: false
    GameId:
      type: object
      properties:
        value:
          type: string
          format: uuid
      additionalProperties: false
    GameState:
      enum:
        - None
        - New
        - Ready
        - Started
        - GameOver
      type: string
    GameStateResponse:
      type: object
      properties:
        state:
          $ref: '#/components/schemas/GameState'
        winner:
          $ref: '#/components/schemas/BoardSide'
      additionalProperties: false
    GetGameQueryResult:
      type: object
      properties:
        gameId:
          type: string
          format: uuid
        playerId:
          type: string
          format: uuid
        boardSize:
          type: integer
          format: int32
        state:
          $ref: '#/components/schemas/GameState'
        winnerSide:
          $ref: '#/components/schemas/BoardSide'
      additionalProperties: false
    LastRoundResult:
      type: object
      properties:
        gameId:
          $ref: '#/components/schemas/GameId'
        playerTargetCell:
          type: string
          nullable: true
        playerAttackResult:
          $ref: '#/components/schemas/CellState'
        playerSunkShip:
          $ref: '#/components/schemas/ShipKind'
        opponentTargetCell:
          type: string
          nullable: true
        opponentAttackResult:
          $ref: '#/components/schemas/CellState'
        opponentSunkShip:
          $ref: '#/components/schemas/ShipKind'
        gameState:
          $ref: '#/components/schemas/GameState'
        winnerSide:
          $ref: '#/components/schemas/BoardSide'
      additionalProperties: false
    OpponentStrategy:
      enum:
        - None
        - Random
        - SemanticKernel
      type: string
    PlaceShipRequest:
      type: object
      properties:
        side:
          $ref: '#/components/schemas/BoardSide'
        shipKind:
          $ref: '#/components/schemas/ShipKind'
        orientation:
          $ref: '#/components/schemas/ShipOrientation'
        bowCode:
          type: string
          nullable: true
      additionalProperties: false
    PlayerResponse:
      type: object
      properties:
        id:
          type: string
          description: The player's unique identifier.
          format: uuid
        username:
          type: string
          description: The player's username.
          nullable: true
        activeGameId:
          type: string
          description: 'The currently active game ID, if any.'
          format: uuid
          nullable: true
        totalGamesPlayed:
          type: integer
          description: The total number of games played.
          format: int32
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
        title:
          type: string
          nullable: true
        status:
          type: integer
          format: int32
          nullable: true
        detail:
          type: string
          nullable: true
        instance:
          type: string
          nullable: true
      additionalProperties: { }
    RefreshRequest:
      type: object
      properties:
        refreshToken:
          type: string
          nullable: true
      additionalProperties: false
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
    ShipKind:
      enum:
        - None
        - Destroyer
        - Cruiser
        - Submarine
        - Battleship
        - Carrier
      type: string
    ShipOrientation:
      enum:
        - None
        - Vertical
        - Horizontal
      type: string
    SignInRequest:
      type: object
      properties:
        username:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
      additionalProperties: false
    UpdateGameStateRequest:
      type: object
      properties:
        state:
          $ref: '#/components/schemas/GameState'
      additionalProperties: false
  securitySchemes:
    Bearer:
      type: http
      description: Paste the access token from POST /api/auth/signin (no 'Bearer ' prefix).
      scheme: bearer
      bearerFormat: JWT
security:
  - Bearer: [ ]
tags:
  - name: Auth
  - name: Games
  - name: Players